==> 01_Extract_Data.ipynb <==
import pandas as pd
import pickle

def extract_data(week_nums):
    url = "http://web.mta.info/developers/data/nyct/turnstile/turnstile_{}.txt"
    df_list = []
    for week in week_nums:
        file_url = url.format(week)
        df = pd.read_csv(file_url)
        df_list.append(df)
    return pd.concat(df_list)
    
    
turnstile_data = extract_data([180630,180623,180616])

with open('mta_extract.pickle', 'wb') as to_write:
    pickle.dump(turnstile_data, to_write)
