==> 02_clean_data.ipynb <==
import pandas as pd
import pickle
import datetime

with open('mta_extract.pickle','rb') as read_file:
    mta_data = pickle.load(read_file)

#Convert DATE to date format
mta_data['DAY'] = pd.to_datetime(mta_data['DATE'])
mta_data.columns = [column.strip() for column in mta_data.columns]

#Get rid of duplicates and Recover Audit entries
by_day = mta_data[(mta_data['DESC'] == 'REGULAR')]

#Create a data frame that looks at each turnstile and identifies the lowest recorded entries for the day. 
#We assume that the first count of the day is the beginning of the day/end of previous day
by_day = by_day.sort_values(['C/A','UNIT','SCP','DAY','ENTRIES'], ascending = True).groupby(['C/A','UNIT','SCP','DAY']).head(1)

by_day.head()

#Set DAY to previous day. This allows us to assign today's lowest count - yesterday's lowest count = yesterday's total entries.
by_day['DAY'] = by_day['DAY'] - pd.Timedelta(1,'D')
by_day['COUNT'] = by_day.sort_values(['C/A','UNIT','SCP','DAY']).groupby(['C/A','UNIT','SCP'])['ENTRIES'].diff()
by_day['COUNT OF EXITS'] = by_day.sort_values(['C/A','UNIT','SCP','DAY']).groupby(['C/A','UNIT','SCP'])['EXITS'].diff()

#get rid of NaN and correct negative counters

by_day.dropna(axis = 0,inplace = True)
by_day = by_day.sort_values(['C/A','UNIT','SCP','LINENAME','DIVISION','STATION','DAY'])

# Create function to correct negative counters and counters that defy logic. i.e. more than 90000 people through the turnstile
def data_cleanse(row, column):
    counter = row[column]
    if counter < 0:
        counter = -counter
    if counter > 90000:
        counter = 0
    return counter

by_day['DAILY COUNTS'] = by_day.apply(data_cleanse, axis = 1, column = 'COUNT')
by_day['DAILY EXITS'] = by_day.apply(data_cleanse, axis = 1, column = 'COUNT OF EXITS')


by_day['NAME OF DAY'] = by_day['DAY'].dt.weekday_name
by_day['WEEK NUMBER'] = by_day['DAY'].dt.week

def fixed_week(row):
    week = row['WEEK NUMBER']
    if row['NAME OF DAY'] == 'Saturday':
        week += 1
    elif row['NAME OF DAY'] == 'Sunday':
        week += 1
    return week

by_day['WEEK NUMBER'] = by_day.apply(fixed_week, axis=1)

by_day = by_day.drop(['TIME','DESC','ENTRIES','EXITS','COUNT','COUNT OF EXITS'], axis = 1, errors = 'ignore')

with open('mta_clean_by_day.pickle', 'wb') as to_write:
    pickle.dump(by_day, to_write)
