==> 03.5_Station_import.ipynb <==
import pandas as pd
import pickle

station_detail = pd.read_csv('alford3.csv', header = None)

station_detail.columns = ['STATION','DIVISION','1','Division','Stop Name','2','3','4']

station_detail.drop(['1','2','3','4'], axis = 1, errors = 'ignore', inplace = True)
station_detail.head()

station_csv = pd.read_csv('http://web.mta.info/developers/data/nyct/subway/Stations.csv')
station_csv.head()

stations = station_csv.sort_values(['Stop Name','Division']).groupby(['Stop Name','Division']).head(1)

mta_station_detail = pd.merge(station_detail,stations, how = 'left', left_on = ['Stop Name','Division'], right_on = ['Stop Name','Division'])

with open('mta_station_detail.pickle', 'wb') as to_write:
    pickle.dump(mta_station_detail, to_write)
