==> 04_Combine_with_latlong.ipynb <==
import pandas as pd
import pickle

with open('mta_station_detail.pickle','rb') as read_file:
    mta_station_detail = pickle.load(read_file)
with open('mta_counts_total.pickle','rb') as read_file:
    mta_total_counts = pickle.load(read_file)

mta_station_detail.head()

mta_total_counts.head(20)

merged_counts = pd.merge(mta_total_counts,mta_station_detail, how = 'left', left_on = ['STATION','DIVISION'],right_on = ['STATION','DIVISION'])

merged_counts.to_csv('counts_with_latlong.csv',sep = '\t')
