==> 05_add_census_data.ipynb <==
import numpy as np
import pandas as pd
import requests
import urllib.request, json
import pickle


latlonglist = pd.read_csv('counts_with_latlong.csv', sep='\t', index_col=0)

#latitude_coords = latlonglist['GTFS Latitude']
latlonglist_nonempty = latlonglist.dropna(subset=['GTFS Latitude', 'GTFS Longitude'] )

set(latlonglist_nonempty['GTFS Longitude'].isnull())

latlonglist_nonempty.reset_index(inplace=True)

FCC_urls = []


for i in range(0, len(latlonglist_nonempty['STATION'])):
    FCC_urls.append('https://geo.fcc.gov/api/census/block/find?latitude={}&longitude={}&format=json'.format(latlonglist_nonempty['GTFS Latitude'][i], latlonglist_nonempty['GTFS Longitude'][i]))


geoid = []

for i in FCC_urls:
    with urllib.request.urlopen(i) as url:
        data = json.loads(url.read().decode())
        geoid.append(data['Block']['FIPS'])


latlonglist_nonempty['GEOID10'] = [int(x[0:11]) for x in geoid]

census_female_pop = pd.read_csv('2010_FEMALE_POPULATION2.csv', sep=',', header=0)

totalcounts_merged_femalepop = pd.merge(latlonglist_nonempty, census_female_pop, on='GEOID10', how='left')
                                                

with open('pop_merged', 'wb') as to_write:
    pickle.dump(totalcounts_merged_femalepop, to_write)

with open('pop_merged', 'rb') as read_file:
    pop_merged = pickle.load(read_file)

more_census_data = pd.read_csv('2010_more_demogs.csv', sep=',')

more_census_data.drop(labels='2010_Female_Population_(US_Census)', axis=1, inplace=True)

census_merged_counts = pd.merge(pop_merged, more_census_data, on='GEOID10', how='left')



with open('census_merged_counts', 'wb') as to_write:
    pickle.dump(census_merged_counts, to_write)
