==> 06_weighted_model.ipynb <==
import pandas as pd
import pickle

with open('census_merged_counts','rb') as read_file:
    census_data = pickle.load(read_file)

model_data = census_data

model_data.head()

list(model_data)

# remove columns from model data that is unnecessary.
to_remove = (
            ['index',
            #'STATION',
             'DIVISION',
             #'DAILY COUNTS',
             'DAILY EXITS',
             'Division',
             'Stop Name',
             'Station ID',
             'Complex ID',
             'GTFS Stop ID',
             #'Line',
             'Borough',
             'Daytime Routes',
             'Structure',
             #'GTFS Latitude',
             #'GTFS Longitude',
            'GEOID10',
             #'2010_Female_Population_(US_Census)',
            'District'
             #'2016_Median_HHIncome',
             #'2016_Education_Bach',
             #'2016_Education_Grad'
            ])
model_data = model_data[model_data.columns.difference(to_remove)] # this reverse column order
model_data = model_data[model_data.columns[::-1]]               # putting columns back

model_data.head()

# Create functions to calculate distance from Gala.

from math import radians, sin, cos, acos
def distancekm(slat,slon,elat,elon):
    #slat = radians(float(input("Starting latitude: ")))
    #slon = radians(float(input("Starting longitude: ")))
    #elat = radians(float(input("Ending latitude: ")))
    #elon = radians(float(input("Ending longitude: ")))
    slat = radians(float(slat))
    slon = radians(float(slon))
    elat = radians(float(elat))
    elon = radians(float(elon))
    dist = (6371.01 * acos(sin(slat)*sin(elat) + cos(slat)*cos(elat)*cos(slon - elon)))
    return dist 
#distance2_points(23.5,67.5,25.5,69.5)
# 300.669463012709 km
def distgala(row):
    #slat = radians(float(input("Starting latitude: ")))
    #slon = radians(float(input("Starting longitude: ")))
    #elat = radians(float(input("Ending latitude: ")))
    #elon = radians(float(input("Ending longitude: ")))
    slat = row['GTFS Latitude']
    slon = row['GTFS Longitude']
    slat = radians(float(slat))
    slon = radians(float(slon))
    elat = radians(float(40.7150))
    elon = radians(float(-73.9843))
    dist = (6371.01 * acos(sin(slat)*sin(elat) + cos(slat)*cos(elat)*cos(slon - elon)))
    return dist 
# LAT LONG LowerEast Side: lat = 40.7150; lon = -73.9843;

model_data['Distance'] = model_data.apply(distgala,axis = 1) 

# remove Lat Long Columns from Data
to_remove = ([
             'GTFS Latitude',
             'GTFS Longitude',
            ])
model_data = model_data[model_data.columns.difference(to_remove)] # this reverse column order
model_data = model_data[model_data.columns[::-1]]               # putting columns back

model_data.head()

model_data2=model_data.rename(index=str, columns={"2016_Median_HHIncome": "Income", "2010_Female_Population_(US_Census)": "W"})


model_data2.head()

Dmin = min(model_data2['Distance'])
CountMax = max(model_data2['DAILY COUNTS'])
IncomMax = max(model_data2['Income'])
GraduMax = max(model_data2['2016_Education_Grad'])
Bachmax = max(model_data2['2016_Education_Bach'])
Wmax = max(model_data2['W'])

# Recommendation Index = RI = Function of (
#		Number of entries per station
#		Distance from Gala location (Lower East Side)
#		% Women for zip code of Station
#		Number of people with Bachelor degree in GEOID of station
#		Number of people with Graduate degree in GEOID of station
#		Medium Household Income in GEOID of station
#				)
# RI(station) = [(Entries/Max Entries) + (Income/Max Income) 
#               + (NumberBachelor/MaxNOB) + (NumberGraduate/MaxNOG)
#               + (NumberOfWomen/MaxNumberWomen) + (DistanceMinimum/Distance)]
#                / 6 (there are 6 factors)
#               * 100 (result between 0 and 100 )

model_data3 = model_data2.dropna()

model_data3.describe()

# Create model based on weighting scale and made Recommenadation Index out of 100 pts.
model_data3['RI'] = (
            (
                  2*(model_data3['DAILY COUNTS'] / CountMax)   
                + 1.3*(model_data3['Income'] / IncomMax)
                + 1.4*(model_data3['2016_Education_Bach'] / Bachmax)
                + 1.1*(model_data3['2016_Education_Grad'] / GraduMax)
                + 3*(Dmin / model_data3['Distance'])
                + 2.5*(model_data3['W'] / Wmax)
            )*100/11.3
                    )

model_data3.sort_values(by='RI', ascending=False).head(10)

model_data3.sort_values(by='', ascending=False).head(10)
